Make linear gradient calculable
