+++ packages/alfa-css/src/value/image/gradient/item/index.ts
@@ -1,0 +1,1 @@
+export * from "./item";
