+++ packages/alfa-css/src/value/image/gradient/linear/index.ts
@@ -1,0 +1,1 @@
+export * from "./linear";
