+++ packages/alfa-css/tsconfig.json
@@ -46,12 +46,18 @@
     "src/value/color/rgb.ts",
     "src/value/color/system.ts",
     "src/value/image/gradient/gradient.ts",
-    "src/value/image/gradient/hint.ts",
     "src/value/image/gradient/index.ts",
-    "src/value/image/gradient/item.ts",
-    "src/value/image/gradient/linear.ts",
-    "src/value/image/gradient/radial.ts",
-    "src/value/image/gradient/stop.ts",
+    "src/value/image/gradient/item/hint.ts",
+    "src/value/image/gradient/item/index.ts",
+    "src/value/image/gradient/item/item.ts",
+    "src/value/image/gradient/item/stop.ts",
+    "src/value/image/gradient/linear/corner.ts",
+    "src/value/image/gradient/linear/direction.ts",
+    "src/value/image/gradient/linear/index.ts",
+    "src/value/image/gradient/linear/linear.ts",
+    "src/value/image/gradient/linear/position.ts",
+    "src/value/image/gradient/linear/side.ts",
+    "src/value/image/gradient/radial/radial.ts",
     "src/value/image/image.ts",
     "src/value/image/index.ts",
     "src/value/image/url.ts",
